requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
tenacity>=8.2.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.31.0
supabase>=2.0.0
python-dateutil>=2.8.2
